feat: 기존의 FrontController 패키지 이동, 찾지 못할시 예외메시지 추가.
