--- src/main/java/controller/FrontController.java
@@ -1,31 +1,0 @@
-package controller;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import request.HttpRequest;
-import response.HttpResponse;
-import util.HttpMethod;
-import webserver.RequestHandler;
-
-import static util.HttpMethod.*;
-
-public abstract class FrontController implements Controller {
-    private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
-
-    @Override
-    public void service(HttpRequest request, HttpResponse response) {
-        HttpMethod method = request.getMethod();
-        log.debug("FrontController method : {}", request.getMethod());
-        if (method.equals(GET)) {
-            doGet(request, response);
-        } else if (method.equals(POST)) {
-            doPost(request, response);
-        }
-    }
-
-    protected void doGet(HttpRequest request, HttpResponse response) {
-    }
-
-    protected void doPost(HttpRequest request, HttpResponse response) {
-    }
-}
